% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{audio_features_summarised}
\alias{audio_features_summarised}
\title{Audio features summarised}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 31 columns.
}
\usage{
audio_features_summarised
}
\description{
Audio features summarised
}
\keyword{datasets}
