% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{audio_features_summarised_by_syllable}
\alias{audio_features_summarised_by_syllable}
\title{Audio features summarised by syllable}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 32 columns.
}
\usage{
audio_features_summarised_by_syllable
}
\description{
Audio features summarised by syllable
}
\keyword{datasets}
