% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_syllables.R
\name{extract_syllables}
\alias{extract_syllables}
\title{Extract syllables from an audio file}
\usage{
extract_syllables(audio_file)
}
\arguments{
\item{audio_file}{}
}
\description{
Extract syllables from an audio file
}
