% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{prepped_recipe}
\alias{prepped_recipe}
\title{A prepped tidymodels recipe for the sylalble classifier model}
\format{
An object of class \code{recipe} of length 10.
}
\usage{
prepped_recipe
}
\description{
A prepped tidymodels recipe for the sylalble classifier model
}
\keyword{datasets}
