% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{shap_values_baseline}
\alias{shap_values_baseline}
\title{Baseline Shapley values}
\format{
An object of class \code{mshapviz} of length 8.
}
\usage{
shap_values_baseline
}
\description{
Baseline Shapley values
}
\keyword{datasets}
