% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{syllable_classifier_bundle}
\alias{syllable_classifier_bundle}
\title{Syllable classifier model}
\format{
An object of class \code{bundled_workflow} (inherits from \code{bundle}) of length 2.
}
\usage{
syllable_classifier_bundle
}
\description{
Syllable classifier model
}
\keyword{datasets}
