% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{vowel_metadata_with_audio_features_range_restricted_without_AEIOU}
\alias{vowel_metadata_with_audio_features_range_restricted_without_AEIOU}
\title{The training dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 307 rows and 47 columns.
}
\usage{
vowel_metadata_with_audio_features_range_restricted_without_AEIOU
}
\description{
The training dataset
}
\keyword{datasets}
